# docs/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GHCAA.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ledger": {
      "get": {
        "tags": [
          "FinancialLedger"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FinancialRecordType"
            }
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "FinancialLedger"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialRecord"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialRecord"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ledger/summary": {
      "get": {
        "tags": [
          "FinancialLedger"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ledger/{id}": {
      "put": {
        "tags": [
          "FinancialLedger"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialRecord"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialRecord"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "FinancialLedger"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ledger/export/csv": {
      "get": {
        "tags": [
          "FinancialLedger"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinancialCategory": {
        "enum": [
          "MembershipFee",
          "RegistrationFee",
          "Donation",
          "Event",
          "Maintenance",
          "Salary",
          "Utilities",
          "ReunionFee",
          "Sponsorship",
          "Grant",
          "Refund",
          "Other"
        ],
        "type": "string"
      },
      "FinancialRecord": {
        "required": [
          "amount",
          "date",
          "description",
          "financialCategory",
          "recordType",
          "year"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "recordType": {
            "$ref": "#/components/schemas/FinancialRecordType"
          },
          "financialCategory": {
            "$ref": "#/components/schemas/FinancialCategory"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "reference": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdByAdminId": {
            "type": "integer",
            "format": "int32"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedByAdminId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedByAdminId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FinancialRecordType": {
        "enum": [
          "Income",
          "Expense"
        ],
        "type": "string"
      },
...
    }
  }
}